# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "description": "Request model for creating a project.",
        "properties": {
          "clientId": {
            "title": "Clientid",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "endDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "medium",
            "title": "Priority"
          },
          "startDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          }
        },
        "required": [
          "name",
          "code",
          "clientId",
          "phase"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse_ProjectResponse_": {
        "example": {
          "items": [],
          "page": 1,
          "per_page": 20,
          "total": 42
        },
        "properties": {
          "items": {
            "description": "List of items for current page",
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "Current page number (1-indexed)",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "description": "Number of items per page",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items available",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "per_page",
          "items"
        ],
        "title": "PaginatedResponse[ProjectResponse]",
        "type": "object"
      },
...
      "ProjectResponse": {
        "description": "Response model for project data.",
        "properties": {
          "clientId": {
            "title": "Clientid",
            "type": "string"
          },
          "clientName": {
            "title": "Clientname",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "completedStories": {
            "default": 0,
            "title": "Completedstories",
            "type": "integer"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "endDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "startDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "storiesCount": {
            "default": 0,
            "title": "Storiescount",
            "type": "integer"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "description",
          "createdBy",
          "clientId",
          "clientName",
          "status",
          "priority",
          "phase",
          "startDate",
          "endDate",
          "createdAt",
          "updatedAt"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
...
      "UpdateProjectRequest": {
        "description": "Request model for updating a project.",
        "properties": {
          "clientId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clientid"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "endDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "startDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          }
        },
        "title": "UpdateProjectRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for Open Projects Hub \u2014 project, client, and story management with JWT authentication and role-based access control. See [docs/api/README.md](https://github.com/alonsovndev/open-projects-hub-api/blob/main/docs/api/README.md) for auth flows, pagination, and code examples.",
    "title": "Open Projects Hub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects": {
      "get": {
        "description": "List projects with pagination and optional filters.\n\nRequires authentication.\n\nArgs:\n    limit: Maximum number of results (1-100, default 20)\n    offset: Number of results to skip (default 0)\n    project_status: Optional status filter (active, completed, archived)\n    client_id: Optional client UUID filter\n    created_from: Optional lower bound on created_at (ISO datetime)\n    created_to: Optional upper bound on created_at (ISO datetime)\n    updated_from: Optional lower bound on updated_at (ISO datetime)\n    updated_to: Optional upper bound on updated_at (ISO datetime)\n    search: Optional substring match on name or code (case-insensitive)\n    current_user: Current authenticated user\n    use_case: Injected ListProjectsUseCase\n\nReturns:\n    PaginatedResponse containing pagination metadata and list of ProjectResponse objects\n\nRaises:\n    400: Invalid query parameters\n    401: Unauthorized\n    500: Internal server error",
        "operationId": "list_projects_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clientid"
            }
          },
          {
            "in": "query",
            "name": "createdFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Createdfrom"
            }
          },
          {
            "in": "query",
            "name": "createdTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Createdto"
            }
          },
          {
            "in": "query",
            "name": "updatedFrom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updatedfrom"
            }
          },
          {
            "in": "query",
            "name": "updatedTo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updatedto"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Create a new project.\n\nRequires ADMIN role. The created_by user is extracted from the JWT token.\n\nArgs:\n    payload: CreateProjectRequest with project details (name, code, client_id, optional fields)\n    current_user: Current authenticated admin user (from JWT)\n    use_case: Injected CreateProjectUseCase\n\nReturns:\n    ProjectResponse with created project data (id, name, status, timestamps)\n\nRaises:\n    400: Validation failed (invalid name, dates, etc.)\n    401/403: Unauthorized or forbidden\n    500: Internal server error",
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/v1/projects/{project_id}": {
      "delete": {
        "description": "Delete project.\n\nRequires ADMIN role. Cascade deletes all related stories.\n\nArgs:\n    project_id: Project UUID\n    current_user: Current authenticated admin user\n    use_case: Injected DeleteProjectUseCase\n\nRaises:\n    400: Invalid UUID\n    401/403: Unauthorized or forbidden\n    404: Project not found\n    500: Internal server error",
        "operationId": "delete_project_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "description": "Get project by ID.\n\nRequires authentication.\n\nArgs:\n    project_id: Project UUID\n    current_user: Current authenticated user\n    use_case: Injected GetProjectByIdUseCase\n\nReturns:\n    ProjectResponse with project data\n\nRaises:\n    400: Invalid UUID\n    401: Unauthorized\n    404: Project not found\n    500: Internal server error",
        "operationId": "get_project_by_id_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Project By Id",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "description": "Update project.\n\nRequires ADMIN role. Supports partial updates \u2014 only provided fields are changed.\n\nArgs:\n    project_id: Project UUID\n    payload: UpdateProjectRequest with fields to update\n    current_user: Current authenticated admin user\n    use_case: Injected UpdateProjectUseCase\n\nReturns:\n    ProjectResponse with updated project data\n\nRaises:\n    400: Validation failed\n    401/403: Unauthorized or forbidden\n    404: Project not found\n    500: Internal server error",
        "operationId": "update_project_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/v1/projects/{project_id}/archive": {
      "post": {
        "description": "Archive a project, excluding it from active-project limits.\n\nRequires ADMIN role.\n\nRaises:\n    404: Project not found\n    401/403: Unauthorized or forbidden",
        "operationId": "archive_project_v1_projects__project_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/v1/projects/{project_id}/reactivate": {
      "post": {
        "description": "Reactivate an archived or completed project.\n\nRequires ADMIN role. Subject to active-project limit enforcement (BE-002).\n\nRaises:\n    404: Project not found\n    409: Active project limit exceeded\n    401/403: Unauthorized or forbidden",
        "operationId": "reactivate_project_v1_projects__project_id__reactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reactivate Project",
        "tags": [
          "Projects"
        ]
      }
    },
...
  }
}